Move type definitions closer to point of use
